requirements.txt
httpx>=0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
APScheduler>=3.10.4
python-dotenv>=1.0.1
lxml==5.3.0
fastapi>=0.115.0
uvicorn>=0.30.6
pywhatkit>=5.4
fpdf2>=2.7.9
Pillow>=10.3.0
